requirements.txt
# Production dependencies (same as dockerfile)
# Note: essentia-tensorflow is not available on PyPI for Windows
# For dev: Skip essentia, most QC tools work without it
# essentia-tensorflow==2.1b6.dev1389
numpy==1.26.4
scipy>=1.13.1  # Unpinned for Windows dev (needs pre-built wheels for newer Python versions)
soundfile
mutagen==1.47.0
pyyaml==6.0.2
fastapi==0.115.0
uvicorn==0.30.6
rich==13.7.1
requests==2.32.3
bcrypt==4.2.1
python-arango==7.9.0

# Testing dependencies
pytest==8.3.3
pytest-cov==5.0.0

# Development/QC tools
ruff>=0.1.0              # Linting and formatting
mypy>=1.7.0              # Static type checking
bandit>=1.7.5            # Security issue scanner
vulture>=2.10            # Dead code detection
pydub>=0.25.1            # Audio file generation for tests (requires ffmpeg)

# Complexity & Architecture Analysis
radon>=5.1.0,<5.2        # Cyclomatic complexity, maintainability index (constrained by wily)
import-linter>=2.0       # Enforce layered architecture rules
wily>=1.25.0             # Track complexity trends over time

# Pattern Detection & Code Smells
flake8>=7.0.0                       # Base linter
flake8-cognitive-complexity>=0.1.0  # Measures code understandability
flake8-simplify>=0.21.0             # Detects overcomplicated patterns
flake8-eradicate>=1.5.0             # Finds commented-out code
flake8-variables-names>=0.0.6       # Generic variable name detection

# Type stubs for better mypy support
types-PyYAML
types-requests
